+++ R/build_tools.R
@@ -57,7 +57,10 @@ verify_version <- function(name,
   ## '>=1.2.3' => '>= 1.2.3'
   required_version <-
     gsub("^([><=]+)([0-9.]+)$", "\\1 \\2", required_version, perl = TRUE)
-  if (!remotes:::version_satisfies_criteria(pkg_version, required_version)) {
+  # TODO: get rid of usage `:::`
+  #nolint start
+  if (!remotes::version_satisfies_criteria(pkg_version, required_version)) {
+  #nolint stop
     stop(sprintf(
       "Invalid version of %s. Installed: %s, required %s.",
       name,