+++ R/linters_config.R
@@ -14,7 +14,6 @@ gDR_undesirable_functions <-
     "read.csv" = "pleae use `data.table::fread` instead (data.table is primary data format)",
     "as.data.frame" = "please use `data.table::as.data.table` instead (data.table is primary data format)",
     "reshape2" = "please use functions from `data.table` package (data.table is primary data format)",
-    # we use logExtra::debug
     "debug" = NULL
   )