+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRstyle 1.5.1 - 2024-11-05
+* synchronize Bioconductor and GitHub versioning
+
 ## gDRstyle 1.3.3 - 2024-08-29
 * add GitLab credentials to pkgdown builds