+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRstyle 1.9.1 - 2025-10-30
+* synchronize Bioconductor and GitHub versioning
+
 ## gDRstyle 1.7.4 - 2025-08-12
 * fix usage of `ifelse`