+++ NEWS.md
@@ -1,5 +1,5 @@
 ## gDRstyle 1.11.4 - 2026-05-18
-* fix linting violations from updated gDRstyle rules
+* apply updated gDRstyle rules
 
 ## gDRstyle 1.11.3 - 2026-05-08
 * extend linter config with additional rules and update style guide accordingly