+++ NEWS.md
@@ -1,4 +1,4 @@
-## gDRstyle 1.11.4 - 2026-05-18
+## gDRstyle 1.11.4 - 2026-05-27
 * apply updated gDRstyle rules
 
 ## gDRstyle 1.11.3 - 2026-05-08